Modify guide17.m to publish properly with the website.
Backticks are special characters in Markdown; using them in the way that
they are in LaTeX causes issues.  LaTeX control sequences like "\vskip 1
em" don't cause problems, but they do get printed when published for the
website.  Finally, for reasons I don't understand, some of the calls to
snapnow() resulted in giant blocks of whitespace in the web version.
Removing them seems to have done the trick.

I recognize that this will make the LaTeX rendering of the document look
worse.  The right thing to do would be to figure out a system to make
things work correcty both for the web and for LaTeX, but I don't have
time to do this right now.